Fixed a mistake in the learning algorithm. Fixed the user deletion issue. Added a possibility to attach pictures to questions and answers.

+++ MemoBoost.Logic/MediaManager.cs
@@ -9,12 +9,41 @@ namespace MemoBoost.Logic
 {
     public class MediaManager
     {
-        public static void Copy(string file)
+        private static string fpath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Media");
+
+        public static void Copy(string file, out string filePath)//worked without out, 
         {
             string mediaFolder = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Media");
             if (!Directory.Exists(mediaFolder)) Directory.CreateDirectory(mediaFolder);
-            string filePath = Path.Combine(mediaFolder, Path.GetFileName(file));
-            File.Copy(file, filePath);
+            filePath = Path.Combine(mediaFolder, Path.GetFileName(file));
+            if (!File.Exists(filePath))
+                File.Copy(file, filePath);
+            else
+            {
+                filePath = Modify(filePath);
+                File.Copy(file, filePath);
+            }
+        }
+
+        private static string Modify(string filepath)
+        {
+            int count = 1;
+            string name = Path.GetFileNameWithoutExtension(filepath);
+            string extension = Path.GetExtension(filepath);
+            string path = Path.GetDirectoryName(filepath);
+            string modified = filepath;
+            while (File.Exists(modified))
+            {
+                string t = string.Format("{0}({1})", name, count++);
+                modified=Path.Combine(path, t + extension);
+            }
+            return modified;
+        }
+
+        public static void Remove(string path)//exception
+        {
+     
+            File.Delete(Path.Combine(fpath, Path.GetFileName(path)));
         }
     }
 }
